Lots of item / item entry updates
DataFragment now requires its model type
ItemEntry now takes the initial set of items as a parameter
DoDataExchange is now async and requires the database context
Fixed item entries not updating in the database

+++ BackpackPlanner/BackpackPlanner/DAL/Models/Personal/PersonalInformation.cs
@@ -175,7 +175,7 @@ namespace EnergonSoftware.BackpackPlanner.DAL.Models.Personal
         public float WeightInUnits
         {
             get { return _settings?.Units.WeightFromGrams(WeightInGrams) ?? WeightInGrams; }
-            set { WeightInGrams = (int)(_settings?.Units.GramsFromWeight(value) ?? WeightInGrams); }
+            set { WeightInGrams = _settings?.Units.GramsFromWeight(value) ?? WeightInGrams; }
         }
 
         private readonly SettingsManager _settingsManager;
